Refactoring on testing.

+++ Services/UserService.cs
@@ -5,8 +5,6 @@ using Entities.Response;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Repositories;
 using Repositories.Extensions;
 using Services.Contracts;
 using Shared.DataTransferObjects;
@@ -18,35 +16,22 @@ namespace Services
 {
     public class UserService : IUserService
     {
-        private readonly ILoggerManager _logger;
         private readonly IMapper _mapper;
         private readonly UserManager<AppUser> _userManager;
-        private readonly IConfiguration _configuration;
-        private readonly IRepositoryManager _repositoryManager;
         private readonly ICloudinaryService _cloudinaryService;
-        private readonly RepositoryContext _repositoryContext;
         private readonly IHttpContextAccessor _httpContextAccessor;
-        const int PHOTO_MAX_ALLOWABLE_SIZE = 1000000;
 
         public UserService
         (
-            ILoggerManager logger,
             IMapper mapper,
             UserManager<AppUser> userManager,
-            IConfiguration configuration,
-            IRepositoryManager repositoryManager,
             ICloudinaryService cloudinaryService,
-            RepositoryContext repositoryContext,
             IHttpContextAccessor httpContextAccessor
         )
         {
-            _logger = logger;
             _mapper = mapper;
             _userManager = userManager;
-            _configuration = configuration;
-            _repositoryManager = repositoryManager;
             _cloudinaryService = cloudinaryService;
-            _repositoryContext = repositoryContext;
             _httpContextAccessor = httpContextAccessor;
         }
 
@@ -60,16 +45,16 @@ namespace Services
             return new ApiOkResponse<DetailedUserToReturnDto>(userToReturn);
         }
 
-        public async Task<ApiBaseResponse> GetDetails(string id)
+        public ApiBaseResponse GetDetails(Guid id)
         {
-            var user = await _userManager.Users
+            var user = _userManager.Users
                 .Include(u => u.UserInformation.Educations)
                 .Include(u => u.UserInformation.WorkExperiences)
                 .Include(u => u.UserInformation.UserSkills)
                 .Include(u => u.UserInformation.Certifications)
                 .Include(u => u.CareerSummary)
                 .Where(u => u.Id.Equals(id))
-                .FirstOrDefaultAsync();
+                .FirstOrDefault();
 
             if (user == null)
                 return new BadRequestResponse(ResponseMessages.UserNotFound);
@@ -77,18 +62,17 @@ namespace Services
             return new ApiOkResponse<ApplicantInformation>(_mapper.Map<ApplicantInformation>(user));
         }
 
-        public async Task<ApiBaseResponse> Get(SearchParameters searchParameters)
+        public ApiBaseResponse Get(SearchParameters searchParameters)
         {
             var endDate = searchParameters.EndDate == DateTime.MaxValue ? searchParameters.EndDate : searchParameters.EndDate.AddDays(1);
-            var users = await _repositoryContext.Users.AsQueryable()
-                                    //.Where(u => u >= searchParameters.StartDate && u.CreatedAt <= endDate)
-                                    .Where(_=> true)
-                                    //.Search(searchParameters.SearchBy)
-                                    //.OrderByDescending(u => u.CreatedAt)
-                                    //.ThenBy(u => u.FirstName)
-                                    .ToListAsync();
+            var users = _userManager.Users.AsQueryable()
+                                    .Where(u => u.CreatedOn >= searchParameters.StartDate && u.CreatedOn <= endDate)
+                                    .Search(searchParameters.SearchBy)
+                                    .OrderByDescending(u => u.CreatedOn)
+                                    .ThenBy(u => u.FirstName)
+                                    .ToList();
 
-            var pagedList = PagedList<AppUser>.ToPagedList(new List<AppUser>(), searchParameters.PageNumber, searchParameters.PageSize);
+            var pagedList = PagedList<AppUser>.ToPagedList(users, searchParameters.PageNumber, searchParameters.PageSize);
 
             var usersDto = _mapper.Map<IEnumerable<DetailedUserToReturnDto>>(pagedList);
             var pagedDataList = PaginatedListDto<DetailedUserToReturnDto>.Paginate(usersDto, pagedList.MetaData);
@@ -196,11 +180,14 @@ namespace Services
             if (uploadResult == null)
                 return new BadRequestResponse(ResponseMessages.PhotoUploadFailed);
 
-            await _cloudinaryService.DeleteFile(user.PublicId);
+            if (!string.IsNullOrEmpty(user.PublicId))
+            {
+                await _cloudinaryService.DeleteFile(user.PublicId);
+            }
 
             user.PhotoUrl = uploadResult.Url;
             user.PublicId = uploadResult.PublicId;
-            user.UpdatedAt = DateTime.Now;
+            user.UpdatedAt = DateTime.UtcNow;
             await _userManager.UpdateAsync(user);
 
             return new ApiOkResponse<string>(ResponseMessages.PhotoUpdateSuccessful);
